requirements.txt
websockets>=12.0
paho-mqtt>=1.6.1
python-dotenv>=0.19.0
